# docs/openapi.yml
openapi: '3.0.2'
info:
  title: Notify API
  version: '1.0'
  description: |
    The following OpenAPI document lists a subset of the available APIs for US Notify.

    We are currently API compatible with GOV.UK Notify. Please visit [their documentation](https://docs.notifications.service.gov.uk/rest-api.html)
    for more information.

    Authorization uses [a JSON Web Token (JWT) bearer token](https://docs.notifications.service.gov.uk/rest-api.html#authorisation-header). The internal-api
    methods use the same scheme, but must use a specific key for the Admin UI.
servers:
  - url: https://notify-api.app.cloud.gov
    description: Production API endpoint
  - url: https://notify-api-staging.app.cloud.gov
    description: Staging API endpoint
  - url: http://localhost:6011
    description: Local development API endpoint
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    serviceObject:
      type: object
      properties:
        active:
          type: boolean
        billing_contact_email_addresses:
          type: string
        billing_contact_names:
          type: string
        billing_reference:
          type: string
        consent_to_research:
          type: string
        contact_link:
          type: string
        count_as_live:
          type: boolean
        created_by:
          type: string
        email_branding:
          type: string
        email_from:
          type: string
        go_live_at:
          type: string
        go_live_user:
          type: string
        id:
          type: string
        inbound_api:
          type: array
        letter_branding:
          type: string
        message_limit:
          type: number
        name:
          type: string
        notes:
          type: string
        organization:
          type: string
        organization_type:
          type: string
          enum: ["federal", "state", "other"]
          default: "federal"
        permissions:
          type: array
          items:
            type: string
        prefix_sms:
          type: boolean
        purchase_order_number:
          type: string
        rate_limit:
          type: number
        research_model:
          type: boolean
        restricted:
          type: boolean
        service_callback_api:
          type: array
        volume_email:
          type: string
        volume_letter:
          type: string
        volume_sms:
          type: string
    userObject:
      type: object
      properties:
        auth_type:
          type: string
        can_use_webauthn:
          type: string
        current_session_id:
          type: string
        email_access_validated_at:
          type: string
        email_address:
          type: string
        failed_login_count:
          type: number
        id:
          type: string
        logged_in_at:
          type: string
        mobile_number:
          type: string
        name:
          type: string
        organizations:
          type: array
          items:
            type: string
        password_changed_at:
          type: string
        permissions:
          type: object
          properties:
            SERVICE_ID:
              type: array
              items:
                type: string
        platform_admin:
          type: boolean
        services:
          type: array
          items:
            type: string
        state:
          type: string
          enum: ["pending", "active", "inactive"]
...
paths:
...
  /user:
    get:
      security:
        - bearerAuth: []
      description: 'Retrieve list of all users'
      tags:
        - internal-api
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/userObject"
...
  /organizations:
    get:
      security:
        - bearerAuth: []
      description: 'Retrieve organization details'
      tags:
        - internal-api
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    active:
                      type: boolean
                    count_of_live_services:
                      type: number
                    domains:
                      type: array
                    id:
                      type: string
                    name:
                      type: string
                    organization_type:
                      type: string
                      enum: ["federal", "state", "other"]
  /service:
    get:
      security:
        - bearerAuth: []
      description: 'Retrieve all services'
      tags:
        - internal-api
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/serviceObject"
...